15-06

+++ controllers/usuarioController.js
@@ -1,14 +1,17 @@
 const bcrypt = require('bcryptjs');
 const jwt = require('jsonwebtoken');
 const Usuario = require('../models/Usuario');
+const { logAction } = require('../controllers/Logger'); // Importe a função logAction
+const Perfil = require('../models/Perfil');
+
 
 // Função para registrar um novo usuário
 exports.registrarUsuario = async (req, res) => {
-  const { nome, email, senha, cargo, cpf } = req.body;
+  const { nome, email, senha, cargo, cpf, perfilId } = req.body;
 
   // Verifique se os campos obrigatórios foram fornecidos
-  if (!nome || !email || !senha || !cargo || !cpf) {
-    return res.status(400).json({ erro: 'Nome, email, senha, cargo e CPF são obrigatórios' });
+  if (!nome || !email || !senha || !cargo || !cpf || !perfilId) {
+    return res.status(400).json({ erro: 'Nome, email, senha, cargo, CPF e perfilId são obrigatórios' });
   }
 
   try {
@@ -37,8 +40,18 @@ exports.registrarUsuario = async (req, res) => {
       dataNascimento: req.body.dataNascimento || null,
       genero: req.body.genero || null,
       endereco: req.body.endereco || null,
-      cpf
+      cpf,
+      perfilId
     });
+
+    // Verifique se req.usuario está definido
+    if (!req.usuario || !req.usuario.id) {
+      console.error('req.usuario ou req.usuario.id está indefinido');
+      return res.status(500).json({ erro: 'Erro interno do servidor' });
+    }
+
+    // Registre a ação no log
+    await logAction(req.usuario.id, 'Registrar Usuário', `Usuário ${novoUsuario.nome} registrado com sucesso`);
 
     res.status(201).json(novoUsuario);
   } catch (error) {
@@ -50,32 +63,30 @@ exports.registrarUsuario = async (req, res) => {
 exports.loginUsuario = async (req, res) => {
   const { email, senha } = req.body;
 
-  // Verifique se o email e a senha foram fornecidos
   if (!email || !senha) {
     return res.status(400).json({ erro: 'Email e senha são obrigatórios' });
   }
 
   try {
-    // Verifique se o usuário com o email fornecido existe no banco de dados
-    const usuario = await Usuario.findOne({ where: { email } });
+    const usuario = await Usuario.findOne({
+      where: { email },
+      include: { model: Perfil }
+    });
     if (!usuario) {
       return res.status(404).json({ erro: 'Usuário não encontrado' });
     }
 
-    // Verifique se a senha fornecida corresponde à senha do usuário no banco de dados
     const senhaCorrespondente = await bcrypt.compare(senha, usuario.senha);
     if (!senhaCorrespondente) {
       return res.status(401).json({ erro: 'Credenciais inválidas' });
     }
 
-    // Se as credenciais estiverem corretas, crie um token de acesso
     const token = jwt.sign(
-      { id: usuario.id, email: usuario.email },
-      process.env.JWT_SECRET, // Você precisa definir essa variável de ambiente
-      { expiresIn: '1h' } // Defina a expiração do token conforme desejado
+      { id: usuario.id, email: usuario.email, perfil: usuario.Perfil.nome },
+      process.env.JWT_SECRET,
+      { expiresIn: '5h' }
     );
 
-    // Envie o token como resposta
     res.status(200).json({ token });
   } catch (error) {
     res.status(500).json({ erro: error.message });
@@ -118,6 +129,10 @@ exports.atualizarUsuario = async (req, res) => {
       res.status(404).json({ erro: 'Usuário não encontrado' });
     } else {
       const usuarioAtualizado = await Usuario.findByPk(id);
+
+      // Registre a ação no log
+      await logAction(req.usuario.id, 'Atualizar Usuário', `Usuário ${usuarioAtualizado.nome} atualizado com sucesso`);
+
       res.status(200).json(usuarioAtualizado);
     }
   } catch (error) {
@@ -128,13 +143,32 @@ exports.atualizarUsuario = async (req, res) => {
 // Função para excluir um usuário pelo ID
 exports.excluirUsuario = async (req, res) => {
   const { id } = req.params;
+  
+  // Verifique se o id foi fornecido
+  if (!id) {
+    return res.status(400).json({ erro: 'ID do usuário não fornecido' });
+  }
+
   try {
-    const usuarioExcluido = await Usuario.destroy({ where: { id } });
-    if (usuarioExcluido === 0) {
-      res.status(404).json({ erro: 'Usuário não encontrado' });
-    } else {
-      res.status(204).end();
+    // Encontre o usuário para garantir que ele existe
+    const usuario = await Usuario.findByPk(id);
+    if (!usuario) {
+      return res.status(404).json({ erro: 'Usuário não encontrado' });
     }
+
+    // Destrua o usuário encontrado
+    await Usuario.destroy({ where: { id } });
+
+    // Verifique se req.usuario está definido
+    if (!req.usuario || !req.usuario.id) {
+      console.error('req.usuario ou req.usuario.id está indefinido');
+      return res.status(500).json({ erro: 'Erro interno do servidor' });
+    }
+
+    // Registre a ação no log
+    await logAction(req.usuario.id, 'Excluir Usuário', `Usuário ID: ${id} excluído com sucesso`);
+
+    res.status(204).end();
   } catch (error) {
     res.status(500).json({ erro: error.message });
   }
